+++ include/MaaToolKit/MaaToolKitAPI.h
@@ -2,5 +2,5 @@
 
 #include "MaaToolKitDef.h"
 
-#include "Device/MaaToolKitDevice.h"
 #include "Config/MaaToolKitConfig.h"
+#include "Device/MaaToolKitDevice.h"
