+++ source/MaaToolKit/AdbDevice/DeviceMgr.cpp
