+++ source/MaaToolKit/AdbDevice/DeviceMgr.h
