+++ source/MaaToolKit/AdbDevice/DeviceMgrMacOS.h
