+++ source/MaaToolKit/AdbDevice/DeviceMgrWin32.h
